ath10k: flatten htt_tx_info
This prepares for unified tx completion handling.
With this we will be able to treat mgmt tx and
data tx almost the same with the exception of
txfrag pointer which is only relevant/present
during data tx.

The main idea is to kill of the htt_tx_info
though it probably is a better idea to do it
step-by-step.

Signed-off-by: Michal Kazior <michal.kazior@tieto.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>