ath10k: make HTC cancellation a generic thing
With this the HIF layer (e.g. PCI) can tell the
upper layers that a packet was aborted and all
related resources must be freed ASAP.

This makes sense for HTT TX in particular. FW
requires us to allocate additional resources for
each transmitted frame and we must not free those
resources until after the whole thing is
considered completed.

This is required to fix a possible memory leak in
PCI CE teardown path.

Signed-off-by: Michal Kazior <michal.kazior@tieto.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>