ath10k: reorder pci shutdown sequences
If a tx completion happened right before a CE
pipe was stopped we'd free the completion related
skbuff. If the skbuff was a part of HTT TX we'd
leak memory and iommu mapping.

This shouldn't happen normally but if FW hangs or
stalls and we unload the driver then we could
memleak.

Signed-off-by: Michal Kazior <michal.kazior@tieto.com>
Signed-off-by: Kalle Valo <kvalo@qca.qualcomm.com>